+++ mobile/library/common/engine_handle.cc
@@ -2,47 +2,36 @@
 
 namespace Envoy {
 
-envoy_status_t EngineHandle::runOnEngineDispatcher(envoy_engine_t,
+envoy_status_t EngineHandle::runOnEngineDispatcher(envoy_engine_t handle,
                                                    std::function<void(Envoy::Engine&)> func) {
-  if (auto e = engine()) {
-    return e->dispatcher().post([func]() {
-      if (auto e = engine()) {
-        func(*e);
-      }
-    });
+  if (auto engine = reinterpret_cast<Envoy::Engine*>(handle)) {
+    return engine->dispatcher().post([engine, func]() { func(*engine); });
   }
   return ENVOY_FAILURE;
 }
 
 envoy_engine_t EngineHandle::initEngine(envoy_engine_callbacks callbacks, envoy_logger logger,
                                         envoy_event_tracker event_tracker) {
-  // TODO(goaway): return new handle once multiple engine support is in place.
-  // https://github.com/envoyproxy/envoy-mobile/issues/332
-  strong_engine_ = std::make_shared<Envoy::Engine>(callbacks, logger, event_tracker);
-  engine_ = strong_engine_;
-  return 1;
+  auto engine = new Envoy::Engine(callbacks, logger, event_tracker);
+  return reinterpret_cast<envoy_engine_t>(engine);
 }
 
-envoy_status_t EngineHandle::runEngine(envoy_engine_t, const char* config, const char* log_level,
-                                       const char* admin_address_path) {
-  // This will change once multiple engine support is in place.
-  // https://github.com/envoyproxy/envoy-mobile/issues/332
-  if (auto e = engine()) {
-    e->run(config, log_level, admin_address_path);
+envoy_status_t EngineHandle::runEngine(envoy_engine_t handle, const char* config,
+                                       const char* log_level, const char* admin_address_path) {
+  if (auto engine = reinterpret_cast<Envoy::Engine*>(handle)) {
+    engine->run(config, log_level, admin_address_path);
     return ENVOY_SUCCESS;
   }
-
   return ENVOY_FAILURE;
 }
 
-void EngineHandle::terminateEngine(envoy_engine_t) {
-  // Reset the primary handle to the engine, but retain it long enough to synchronously terminate.
-  auto e = strong_engine_;
-  strong_engine_.reset();
-  e->terminate();
+void EngineHandle::terminateEngine(envoy_engine_t handle, bool release) {
+  auto engine = reinterpret_cast<Envoy::Engine*>(handle);
+  engine->terminate();
+  if (release) {
+    // TODO(jpsim): Always delete engine to avoid leaking it
+    delete engine;
+  }
 }
-
-EngineSharedPtr EngineHandle::strong_engine_;
-EngineWeakPtr EngineHandle::engine_;
 
 } // namespace Envoy