+++ mobile/test/java/org/chromium/net/CronetUrlRequestContextTest.java
@@ -1138,8 +1138,10 @@ public class CronetUrlRequestContextTest {
   @Test
   @SmallTest
   @Feature({"Cronet"})
-  @Ignore("Concurrent Engines not yet: https://github.com/envoyproxy/envoy-mobile/issues/2003")
-  public void testInitTwoEnginesSimultaneously() throws Exception {
+  @Ignore(
+      "Multiple Engines are not yet supported: https://github.com/envoyproxy/envoy-mobile/issues/332")
+  public void
+  testInitTwoEnginesSimultaneously() throws Exception {
     // Threads will block on runBlocker to ensure simultaneous execution.
     ConditionVariable runBlocker = new ConditionVariable(false);
     RequestThread thread1 = new RequestThread(mUrl, runBlocker);
@@ -1173,8 +1175,10 @@ public class CronetUrlRequestContextTest {
   @Test
   @SmallTest
   @Feature({"Cronet"})
-  @Ignore("Concurrent Engines not yet: https://github.com/envoyproxy/envoy-mobile/issues/2003")
-  public void testInitDifferentEngines() throws Exception {
+  @Ignore(
+      "Multiple Engines are not yet supported: https://github.com/envoyproxy/envoy-mobile/issues/332")
+  public void
+  testInitDifferentEngines() throws Exception {
     // Test that concurrently instantiating Cronet context's upon various
     // different versions of the same Android Context does not cause crashes
     // like crbug.com/453845