+++ mobile/test/java/org/chromium/net/testing/CronetTestRule.java
@@ -150,6 +150,7 @@ public final class CronetTestRule implements TestRule {
           // Run with the default HttpURLConnection implementation first.
           setTestingSystemHttpURLConnection(true);
           base.evaluate();
+          tearDown();
           // Use Cronet's implementation, and run the same test.
           setTestingSystemHttpURLConnection(false);
           base.evaluate();